Deduplicate HF handler preambles and split repo handlers by domain
Every HF endpoint repeated the same permission-check / ResolvePath / openRepo
/ pre-receive blocks (~15 copies). Extract checkPermission, resolveRepoPath,
openRepoChecked, openRepoDirect, and checkPreReceive helpers that write the
HTTP error responses, and log permission hook failures server-side.

handler_hf_repo.go (900+ lines) is split by domain: info stays, CRUD moves to
handler_hf_repo_crud.go, branch/tag/refs to handler_hf_repo_refs.go, and
commits/compare/super-squash to handler_hf_repo_history.go. Pure moves on top
of the helper rewrite; delete/move/squash keep opening repositories without
the pre-open hook so they never trigger a mirror sync.
